RD-4332 Replace GET /secrets/{key} v3 with v3.1 + nicer error
