+++ rest-service/manager_rest/rest/resources_v3_1/__init__.py
@@ -28,7 +28,6 @@ from .manager_config import (                    # NOQA
 )
 
 from .secrets import (                           # NOQA
-    SecretsKey,
     SecretsExport,
     SecretsImport,
     SecretsSetGlobal,
